# docs/api-contract/openapi.yaml
openapi: 3.0.3
info:
  title: hajaCheck API
  version: "0.38.0-draft"
  description: >
    Contract-First 초안 (PRD §6, §8). 실제 소스 오브 트루스는 이 파일.


    ## 실시간(WS) 상담 — /ws STOMP (FR-7, #20/HAJA-33, #1002/HAJA-496)

    OpenAPI 3.0 은 WebSocket/STOMP 를 네이티브로 표현하지 못하므로, 이 계약은 `paths` 항목이 아니라
    이 서술로 대체한다(`backend/.../counsel/websocket/WebSocketConfig.java` 실구현 기준).

    - **엔드포인트**: `/ws` (SockJS 미사용 — dev(Vite)/prod(nginx) 동일 오리진 프록시 전제, 기본 same-origin
      정책 유지).
    - **인증**: SESSION 쿠키 기반, 2단계로 강제한다 — ① 핸드셰이크(`CounselHandshakeInterceptor`)에서
      SESSION 쿠키 검증, ② 인바운드 채널(`StompAuthChannelInterceptor`)에서 CONNECT 재검증 + SUBSCRIBE
      시점 당사자(참여자) 검증. 클라이언트 페이로드의 senderId 는 신뢰하지 않고 핸드셰이크에서 확립한
      `StompUserPrincipal` 로만 발신자를 식별한다.
    - **destinations**:
      - `SEND /app/counsel/{ticketId}/send` — 상담 메시지 전송(`ChatMessageSendRequest`: content
        ≤2000자/attachmentKey ≤500자, 텍스트·첨부 중 최소 하나). 티켓이 IN_PROGRESS 가 아니거나 발신자가
        참여자(사용자 본인/담당 상담원)가 아니면 예외 대신 조용히 드롭한다(WS 프레임 처리 중 예외로 세션이
        끊기는 것을 피하기 위한 방어적 설계).
      - `SUBSCRIBE /topic/counsel/{ticketId}` — 해당 상담방 브로드캐스트 구독. 서버가 저장된 메시지를
        `ChatMessageResponse` 로 이 destination 에 `convertAndSend` 한다.
      - `SUBSCRIBE /user/queue/counsel/assigned` — 개인 알림(user destination prefix `/user`). 내 티켓이
        상담원에게 배정되면(`POST /api/counsel/tickets/{id}/assign` 성공 시) `CounselTicketResponse` 를
        수신한다.
      - `SUBSCRIBE /user/queue/counsel/ended` — 개인 알림. 내 티켓이 종료되면(`POST /api/counsel/tickets/
...


...


...


...


...


...


...


...


...

...

...

...
servers:
...
paths:
...
  /api/counsel/tickets:
    post:
      summary: 상담 티켓 생성
      description: >
        FR-7(#20/HAJA-33). 시나리오 리프(leadsToCounselor=true)에서 상담원 연결을 요청하면 WAITING 상태의
        티켓을 생성한다. `has_counselor_access=true` 활성 플랜 게이트를 통과해야 하며(회사 소속이면 유효한
        승인 멤버십까지 확인), category/title 은 시나리오 트리를 타고 올라가 스냅샷으로 저장한다(트리 변경과
        무관하게 이력 고정). `CounselTicketController.createTicket` 구현됨.
      security:
        - SessionCookie: []
          XsrfToken: []
      tags: [counsel]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CounselTicketCreateRequest" }
      responses:
        "201":
          description: 생성 성공
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CounselTicketApiResponse" }
        "400":
          description: 유효하지 않은 scenarioId 형식 (INVALID_INPUT)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorApiResponse" }
        "401":
          description: 미인증 (UNAUTHORIZED)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorApiResponse" }
        "403":
          description: >
            has_counselor_access 플랜 미보유/미승인 멤버십 (COUNSEL_PLAN_REQUIRED) 또는
            leadsToCounselor=false 리프로 진입 (COUNSEL_TICKET_FORBIDDEN)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorApiResponse" }
        "404":
          description: 시나리오 노드 없음 (COUNSEL_SCENARIO_NOT_FOUND)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorApiResponse" }
        "500":
          description: >
            category→counselType 매핑 누락 (COUNSEL_TYPE_MAPPING_NOT_FOUND) 또는 참조 Plan 행 없음
            (PLAN_DATA_INVALID) — 둘 다 데이터 정합성 오류.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorApiResponse" }

    get:
      summary: 상담 대기열 조회
      description: >
        FR-7(#20/HAJA-33). 상태별 티켓 목록(생성순 FIFO, 기본 WAITING), 페이지네이션. **COUNSELOR/
        PLATFORM_ADMIN 전용**(SecurityConfig requestMatchers — 대기열 조회는 role 게이트로 엔드포인트
        레벨에서 강제). `CounselTicketController.getQueue` 구현됨.
      security:
        - SessionCookie: []
      tags: [counsel]
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/CounselTicketStatus"
            default: WAITING
        - name: page
          in: query
          schema: { type: integer, default: 0, minimum: 0 }
        - name: size
          in: query
          schema: { type: integer, default: 20, minimum: 1, maximum: 100 }
      responses:
        "200":
          description: 조회 성공
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CounselTicketSummaryPageApiResponse" }
        "401":
          description: 미인증 (UNAUTHORIZED)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorApiResponse" }
        "403":
          description: COUNSELOR/PLATFORM_ADMIN 아님 (FORBIDDEN)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorApiResponse" }
  /api/counsel/tickets/mine:
    get:
      summary: 내 상담 이력 조회
      description: >
        FR-7(#20/HAJA-33). 요청자 본인의 상담 티켓 목록(최신순). `status=ALL`(기본) 또는 특정 상태 필터.
        userId 는 인증 주체(세션)에서만 취득하며 요청 파라미터로 받지 않는다(cross-owner IDOR 방지).
        `CounselTicketController.getMyTickets` 구현됨.
      security:
        - SessionCookie: []
      tags: [counsel]
      parameters:
        - name: status
          in: query
          description: "\"ALL\"(기본, 필터 없음) 또는 CounselTicketStatus 값"
          schema: { type: string, default: ALL }
        - name: page
          in: query
          schema: { type: integer, default: 0, minimum: 0 }
        - name: size
          in: query
          schema: { type: integer, default: 20, minimum: 1, maximum: 100 }
      responses:
        "200":
          description: 조회 성공
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CounselTicketSummaryPageApiResponse" }
        "400":
          description: status 가 ALL 도, 유효한 CounselTicketStatus 도 아님 (INVALID_INPUT)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorApiResponse" }
        "401":
          description: 미인증 (UNAUTHORIZED)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorApiResponse" }
  /api/counsel/tickets/admin:
    get:
      summary: 플랫폼 관리자 날짜별 상담 목록
      description: >
        #1168. 접수일(`createdAt`) 기준 특정 날짜에 접수된 상담 티켓 전체를 최신순으로 반환한다
        (고객 프로필 4필드 포함). **PLATFORM_ADMIN 전용** — 컨트롤러에서 role 을 확인해 그 외 역할은
        COUNSEL_TICKET_FORBIDDEN(403). 날짜 경계는 서버 JVM 기본 타임존 기준
        `[date 00:00:00.000000, date 23:59:59.999999]`(다음날 자정 -1μs) 폐구간으로 계산한다 — PG
        timestamp 가 마이크로초 정밀도라 -1ns 로 끊으면 저장 시 자정으로 올림돼 다음날 티켓이 섞인다.
        `CounselTicketController.getAdminTicketsByDate` 구현됨.
      security:
        - SessionCookie: []
      tags: [counsel]
      parameters:
        - name: date
          in: query
          required: true
          description: "조회할 접수일(YYYY-MM-DD). 누락·형식 오류 모두 400 INVALID_INPUT."
          schema: { type: string, format: date, example: "2026-07-30" }
        - name: page
          in: query
          schema: { type: integer, default: 0, minimum: 0 }
        - name: size
          in: query
          description: "상한 100 — 초과 시 100으로 클램프(거부 아님). 0 이하는 기본값 20."
          schema: { type: integer, default: 20, minimum: 1, maximum: 100 }
      responses:
        "200":
          description: >
            조회 성공. `content[]` 는 CounselTicketSummaryResponse 이며 이 경로에서만 고객 프로필 4필드
            (customerName/customerEmail/customerPlan/customerJoinedAt)가 채워진다.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CounselTicketSummaryPageApiResponse" }
        "400":
          description: date 누락 또는 YYYY-MM-DD 형식 아님 (INVALID_INPUT)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorApiResponse" }
        "401":
          description: 미인증 (UNAUTHORIZED)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorApiResponse" }
        "403":
          description: PLATFORM_ADMIN 아님 (COUNSEL_TICKET_FORBIDDEN)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorApiResponse" }
  /api/counsel/tickets/{id}/messages:
    get:
      summary: 상담 대화 조회
      description: >
        FR-7(#20/HAJA-33). 티켓의 전체 대화 이력(시간순). 당사자(사용자 본인/담당 상담원) 또는
        PLATFORM_ADMIN — 비당사자·미존재 모두 열거 방지를 위해 통일된 404로 응답한다.
        (PLATFORM_ADMIN 우회는 #1168/PR #1205 에서 관리자 콘솔 드릴다운을 위해 추가됐다.)
        `CounselTicketController.getMessages` 구현됨.
      security:
        - SessionCookie: []
      tags: [counsel]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "200":
          description: 조회 성공(대화 없으면 빈 배열)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ChatMessageListApiResponse" }
        "401":
          description: 미인증 (UNAUTHORIZED)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorApiResponse" }
        "404":
          description: 미존재 또는 비당사자 (COUNSEL_TICKET_NOT_FOUND)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorApiResponse" }
  /api/counsel/tickets/{id}/export:
    get:
      summary: 대화 내보내기
      description: >
        FR-7(#20/HAJA-33). 티켓의 전체 대화를 평문 텍스트(.txt, UTF-8)로 다운로드한다. 당사자
        (사용자 본인/담당 상담원)만 — 비당사자·미존재 모두 통일된 404. 응답은 JSON envelope 이 아니라
        `Content-Disposition: attachment` 가 붙은 원시 바이트 스트림이다. `CounselTicketController.
        exportTranscript` 구현됨.
      security:
        - SessionCookie: []
      tags: [counsel]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "200":
          description: 다운로드 성공
          headers:
            Content-Disposition:
              schema: { type: string, example: 'attachment; filename="C-20260727-1.txt"' }
          content:
            text/plain:
              schema: { type: string, format: binary }
        "401":
          description: 미인증 (UNAUTHORIZED)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorApiResponse" }
        "404":
          description: 미존재 또는 비당사자 (COUNSEL_TICKET_NOT_FOUND)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorApiResponse" }
  /api/counsel/tickets/{id}/assign:
    post:
      summary: 상담 티켓 셀프-클레임 배정
      description: >
        FR-7(#20/HAJA-33). 대기(WAITING) 티켓에 자기 자신을 배정하고 COUNSEL 세션을 생성한다(자동 push-배정
        대신 상담원이 대기열에서 직접 집는 pull 모델). **COUNSELOR/PLATFORM_ADMIN 전용**. 스킬 기반 라우팅
        (#743/#772)으로 `counselor_skills`에 티켓 counselType 처리 자격이 없으면 거부한다. 동시 클레임 경합은
        `@Version` 낙관적 락이 감지해 한쪽만 성공하고 다른 쪽은 409(COUNSEL_SESSION_ASSIGNMENT_CONFLICT)로
        표면화한다 — 티켓이 이미 WAITING 이 아닌 경우도 동일 코드. 성공 시 티켓 사용자에게
        `/user/queue/counsel/assigned` 로 실시간 알림을 보낸다(§ WS 서술 참조).
        `CounselTicketController.assign` 구현됨.
      security:
        - SessionCookie: []
          XsrfToken: []
      tags: [counsel]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "200":
          description: 배정 성공
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CounselTicketApiResponse" }
        "401":
          description: 미인증 (UNAUTHORIZED)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorApiResponse" }
        "403":
          description: >
            COUNSELOR/PLATFORM_ADMIN 아님 (FORBIDDEN), 상담사 미존재 (COUNSEL_TICKET_FORBIDDEN) 또는
            해당 counselType 처리 자격 없음 (COUNSEL_SKILL_MISMATCH)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorApiResponse" }
        "404":
          description: 티켓 없음 (COUNSEL_TICKET_NOT_FOUND)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorApiResponse" }
        "409":
          description: >
            이미 WAITING 이 아니거나(선점됨) 동시 클레임 경합으로 낙관적 락 충돌
            (COUNSEL_SESSION_ASSIGNMENT_CONFLICT) — 클라이언트는 대기열을 새로고침해 재시도한다.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorApiResponse" }
  /api/counsel/tickets/{id}/resolve:
    post:
      summary: 상담 종료
      description: >
        FR-7(#20/HAJA-33). 담당 상담원 본인 또는 PLATFORM_ADMIN 이 상담을 종료한다(RESOLVED 전이 + 세션
        종료). 성공 시 티켓 사용자에게 `/user/queue/counsel/ended` 로 실시간 알림을 보낸다(§ WS 서술 참조).
        `CounselTicketController.resolve` 구현됨.
      security:
        - SessionCookie: []
          XsrfToken: []
      tags: [counsel]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "200":
          description: 종료 성공
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CounselTicketApiResponse" }
        "401":
          description: 미인증 (UNAUTHORIZED)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorApiResponse" }
        "403":
          description: 담당 상담원 본인도 PLATFORM_ADMIN 도 아님 (COUNSEL_TICKET_FORBIDDEN)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorApiResponse" }
        "404":
          description: 티켓 없음 (COUNSEL_TICKET_NOT_FOUND)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorApiResponse" }
  /api/counsel/tickets/{id}/leave-offline:
    post:
      summary: 오프라인 이탈
      description: >
        FR-7(#20/HAJA-33). 티켓 소유 사용자 본인만 허용 — WAITING/IN_PROGRESS 티켓을 OFFLINE_LEFT 로
        전이하고 세션을 종료한다. 타인 소유 티켓은 미존재와 통일된 404(cross-user IDOR 열거 방지).
        `CounselTicketController.leaveOffline` 구현됨.
      security:
        - SessionCookie: []
          XsrfToken: []
      tags: [counsel]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "200":
          description: 이탈 처리 성공
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CounselTicketApiResponse" }
        "401":
          description: 미인증 (UNAUTHORIZED)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorApiResponse" }
        "404":
          description: 미존재 또는 타인 소유 (COUNSEL_TICKET_NOT_FOUND)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorApiResponse" }
  /api/counsel/tickets/{id}/note:
    get:
      summary: 상담원 비공개 메모 조회
      description: >
        #1021/HAJA-503. 고객에게 노출되지 않는 상담원 전용 메모(티켓당 1개)를 조회한다. **담당 상담원
        본인만** — 요청자가 티켓의 counselorId 와 다르거나 아직 아무에게도 배정되지 않은 티켓
        (counselorId=null)이면 403(COUNSEL_TICKET_FORBIDDEN). 아직 메모가 없으면 200 + content=null
        (빈 메모)로 응답한다(404 아님). SecurityConfig 의 role 게이트는 COUNSELOR/PLATFORM_ADMIN 이지만
        실질 인가는 서비스의 담당자 검증이므로 PLATFORM_ADMIN 도 담당 상담원이 아니면 403이다.
        `CounselTicketController.getNote` 구현됨.
      security:
        - SessionCookie: []
      tags: [counsel]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "200":
          description: 조회 성공(메모 없으면 content=null)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CounselTicketNoteApiResponse" }
        "401":
          description: 미인증 (UNAUTHORIZED)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorApiResponse" }
        "403":
          description: >
            COUNSELOR/PLATFORM_ADMIN 아님 (FORBIDDEN) 또는 담당 상담원 본인 아님/미배정 티켓
            (COUNSEL_TICKET_FORBIDDEN)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorApiResponse" }
        "404":
          description: 티켓 없음 (COUNSEL_TICKET_NOT_FOUND)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorApiResponse" }
    put:
      summary: 상담원 비공개 메모 저장
      description: >
        #1021/HAJA-503. 상담원 전용 메모를 저장한다 — upsert(없으면 생성, 있으면 갱신). **담당 상담원
        본인만**(403 규칙은 조회와 동일). content 는 null/빈 문자열도 허용해 메모 초기화에 쓰며, 4000자를
        넘으면 400. `CounselTicketController.saveNote` 구현됨.
      security:
        - SessionCookie: []
          XsrfToken: []
      tags: [counsel]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CounselTicketNoteUpdateRequest" }
      responses:
        "200":
          description: 저장 성공
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CounselTicketNoteApiResponse" }
        "400":
          description: content 4000자 초과 (VALIDATION_ERROR)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorApiResponse" }
        "401":
          description: 미인증 (UNAUTHORIZED)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorApiResponse" }
        "403":
          description: >
            COUNSELOR/PLATFORM_ADMIN 아님 (FORBIDDEN) 또는 담당 상담원 본인 아님/미배정 티켓
            (COUNSEL_TICKET_FORBIDDEN)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorApiResponse" }
        "404":
          description: 티켓 없음 (COUNSEL_TICKET_NOT_FOUND)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorApiResponse" }
components:
  securitySchemes:
    InternalKey:
      type: apiKey
      in: header
      name: X-Internal-Key
      description: Spring Boot가 FastAPI 내부 호출에 부착하는 AI_INTERNAL_KEY
    InternalServiceToken:
      type: apiKey
      in: header
      name: X-Internal-Service-Token
      description: Spring Boot와 FastAPI 환경변수에만 주입하는 내부 호출 토큰
    SessionCookie:
      type: apiKey
      in: cookie
      name: SESSION
      description: Spring Security 인증 세션 쿠키
    XsrfToken:
      type: apiKey
      in: header
      name: X-XSRF-TOKEN
      description: 상태 변경 요청에 필요한 double-submit CSRF 토큰
  schemas:
...
    ApiError:
      type: object
      properties:
        code: { type: string, example: AUTH_INVALID_CREDENTIALS }
        message: { type: string, example: 아이디 또는 비밀번호가 올바르지 않습니다. }
...
    ErrorApiResponse:
      type: object
      description: 실패 envelope — error=ApiError
      properties:
        success: { type: boolean, example: false }
        data: { type: object, nullable: true, example: null }
        error: { $ref: "#/components/schemas/ApiError" }

    # ── counsel 도메인 (FR-7, #20/HAJA-33, #1002/HAJA-496) ──
    CounselTicketStatus:
      type: string
      enum: [WAITING, IN_PROGRESS, RESOLVED, OFFLINE_LEFT]
...
    CounselTicketCreateRequest:
      type: object
      required: [scenarioId]
      properties:
        scenarioId:
          type: integer
          format: int64
          description: 상담원 연결을 유발한 시나리오 리프(leadsToCounselor=true) 노드 id
          example: 42
    CounselTicketResponse:
      type: object
      description: 상담 티켓 단건 응답 — 생성/배정/종료 결과 및 개인 알림(/user/queue/counsel/**) 페이로드.
      properties:
        id: { type: integer, format: int64, example: 501 }
        ticketNumber: { type: string, example: "C-20260727-501" }
        category: { type: string, example: INSPECTION_REPORT }
        title: { type: string, example: 분석 결과 문의 }
        userId: { type: integer, format: int64, example: 10 }
        counselorId: { type: integer, format: int64, nullable: true, example: 7 }
        counselorName: { type: string, nullable: true, description: 배정된 상담원 표시 이름(미배정 시 null), example: 김상담 }
        sessionId: { type: integer, format: int64, nullable: true, example: 88 }
        status: { $ref: "#/components/schemas/CounselTicketStatus" }
        queuePosition: { type: integer, nullable: true, example: 3 }
        createdAt: { type: string, format: date-time, example: "2026-07-27T10:00:00" }
        endedAt: { type: string, format: date-time, nullable: true, example: null }
    CounselTicketApiResponse:
      type: object
      description: 성공 envelope — data=CounselTicketResponse
      properties:
        success: { type: boolean, example: true }
        data: { $ref: "#/components/schemas/CounselTicketResponse" }
        error: { type: object, nullable: true, example: null }
...
    CounselTicketSummaryPage:
      type: object
      description: >
        global.common.PageResponse<CounselTicketSummaryResponse> 매핑(content/page/totalElements
        3필드 — size/totalPages 는 이 envelope 에 없음).
      required: [content, page, totalElements]
      properties:
        content:
          type: array
          items: { $ref: "#/components/schemas/CounselTicketSummaryResponse" }
        page: { type: integer, minimum: 0, example: 0 }
        totalElements: { type: integer, format: int64, minimum: 0, example: 12 }
    CounselTicketSummaryPageApiResponse:
      type: object
      description: 성공 envelope — data=CounselTicketSummaryPage
      properties:
        success: { type: boolean, example: true }
        data: { $ref: "#/components/schemas/CounselTicketSummaryPage" }
        error: { type: object, nullable: true, example: null }
    CounselTicketNoteUpdateRequest:
      type: object
      description: >
        상담원 전용 비공개 메모 저장(upsert) 요청. content 는 선택 — null/빈 문자열이면 메모를 비운다.
      properties:
        content:
          type: string
          nullable: true
          maxLength: 4000
          example: 고객 재문의 예정 — 다음 주 점검 보고서 재발송 필요
...
    CounselTicketNoteApiResponse:
      type: object
      description: 성공 envelope — data=CounselTicketNoteResponse
      properties:
        success: { type: boolean, example: true }
        data: { $ref: "#/components/schemas/CounselTicketNoteResponse" }
        error: { type: object, nullable: true, example: null }
...
    ChatMessageListApiResponse:
      type: object
      description: 성공 envelope — data=ChatMessageResponse[]
      properties:
        success: { type: boolean, example: true }
        data:
          type: array
          items: { $ref: "#/components/schemas/ChatMessageResponse" }
        error: { type: object, nullable: true, example: null }
...